+++ dev/02_dev.R
@@ -17,6 +17,7 @@ golem::add_module( name = "count_weixin" ) # Name of the module
 golem::add_module( name = "chiny_trend" ) # Name of the module
 golem::add_module( name = "city_map" ) # Name of the module
 golem::add_module( name = "dxy_table" ) # Name of the module
+golem::add_module( name = "geo_lines" ) # Name of the module
 
 ## 2.2 Add dependencies