+++ inst/app/www/script.js
@@ -1,7 +1,8 @@
 $(document).on('shiny:disconnected', function(event) {
   var dv = document.createElement("DIV"); 
 
-  dv.innerHTML = "<h3>Whoops!You were disconnected</h3>" + 
+  dv.innerHTML = "<h1>Whoops!</h1>" + 
+    "<h3>You were disconnected</h3>" + 
     "<br/>" + 
     "<button class='btn btn-default' onclick='location.reload();'>Reload</button>"